+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Set Maybe type as readonly struct.
+- Added `Maybe::DoAsync` extension overloads.
 
 ## [6.0.0-preview.1] - 2022-03-29